补齐单元测试,测试通过

+++ kiteq-remoting/src/main/java/org/kiteq/remoting/client/KiteIOClient.java
@@ -23,7 +23,7 @@ public interface KiteIOClient {
      * 下一次重连时间秒数
      * @return
      */
-    long getReconnectCount();
+    int getReconnectCount();
 
     boolean reconnect();
